fix warning

+++ Diskcached/Diskcached.m
@@ -52,7 +52,7 @@
     const char *charString = [self UTF8String];
 
     unsigned char result[16];
-    CC_MD5(charString, strlen(charString), result);
+    CC_MD5(charString, (unsigned int)strlen(charString), result);
 
     return [NSString stringWithFormat:@"%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X",
             result[0],  result[1],  result[2],  result[3],
